Read input words and assigns a uniqye id to each.
- create a new HashMap to store our mappings, and a count variable to assign unique IDs.

- get a handle to standard input using io::stdin(), and then create a BufReader around it to efficiently read lines.

- loop over each line of input using stdin.lock().lines().
For each line, we check if it's equal to "-", which is our termination character. If it is, we break out of the loop.
Otherwise, we split the line into words using split_whitespace().

For each word, we use the entry() method of the HashMap to either insert a new key-value pair (using or_insert_with() to generate a unique ID), or update the existing value.

Finally, we print out the contents of the HashMap using println!("{:#?}", map).

Note that the to_owned() method is used to convert the &str slice returned by split_whitespace() into an owned String, since HashMap keys must be owned values.