Build a Hypergraph from each line read.
Each line must be interpreted as an edge.

We define a new struct Hypergraph to represent our hypergraph. It contains two fields: a HashMap of vertices, where each vertex name maps to a set of line numbers where that vertex appears, and a Vec of edges, where each edge is a set of vertex names.

We define methods on the Hypergraph struct to add vertices and edges.

We create a new Hypergraph instance, a line number counter, and a handle to standard input using io::stdin().

We loop over each line of input using stdin.lock().lines().

For each line, we check if it's equal to "-", which is our termination character. If it is, we break out of the loop.

Otherwise, we create a new HashSet to store the vertices for this line's edge, and loop over each word in the line using split_whitespace().

For each word, we add it as a vertex to the Hypergraph, and insert it into our vertices set.

After processing all the words in the line, we add the vertices set as a new edge to the Hypergraph.

We increment the line_num counter to keep track of which line we're on.

Finally, we print out the contents of the Hypergraph using println!("{:#?}", hypergraph).

The resulting Hypergraph contains information about which vertices appear on each line, and which vertices are connected by each edge.